+++ src/io/Joypad.java
@@ -1,4 +1,4 @@
-package nes;
+package io;
 
 /**
  * Captures the logic to read from an input controller
