+++ src/io/NoopController.java
@@ -1,10 +1,10 @@
-package nes;
+package io;
 
 /**
  * This is a dummy class that is used by VIEWER clients in the web application (because we don't need to actually listen
  * for their button presses).
  */
-public class NoopController implements Joypad {
+public class NoopController implements Joypad, Controller {
     @Override
     public byte read() {
         return 0x0;
@@ -12,5 +12,9 @@ public class NoopController implements Joypad {
 
     @Override
     public void write(byte value) {
+    }
+
+    @Override
+    public void initializeListener() {
     }
 }
