+++ src/nes/NESLogComparisonTest.java
@@ -1,5 +1,7 @@
 package nes;
 
+import io.KeyboardController;
+
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileNotFoundException;
